Extracted out ConfChange.task to changeDeckConfiguration function
